refactor: Viewer - move getViewerComponentName to ViewerByFile
